Kernel: Xiaomi kernel changes for REDMI Note 15 SE 5G / REDMI Note 15 Android V
The Patch based on QualComm release TAG: lc-snapdragon_high_mid_2024.SPF.1.0_r1.0.r1_00020.1
The kernel config file used is kunzite_consolidate.config&kunzite_perf.config

+++ tools/mm/page-types.c
@@ -22,7 +22,6 @@
 #include <time.h>
 #include <setjmp.h>
 #include <signal.h>
-#include <inttypes.h>
 #include <sys/types.h>
 #include <sys/errno.h>
 #include <sys/fcntl.h>
@@ -393,9 +392,9 @@ static void show_page_range(unsigned long voffset, unsigned long offset,
 		if (opt_file)
 			printf("%lx\t", voff);
 		if (opt_list_cgroup)
-			printf("@%" PRIu64 "\t", cgroup0);
+			printf("@%llu\t", (unsigned long long)cgroup0);
 		if (opt_list_mapcnt)
-			printf("%" PRIu64 "\t", mapcnt0);
+			printf("%lu\t", mapcnt0);
 		printf("%lx\t%lx\t%s\n",
 				index, count, page_flag_name(flags0));
 	}
@@ -421,9 +420,9 @@ static void show_page(unsigned long voffset, unsigned long offset,
 	if (opt_file)
 		printf("%lx\t", voffset);
 	if (opt_list_cgroup)
-		printf("@%" PRIu64 "\t", cgroup);
+		printf("@%llu\t", (unsigned long long)cgroup);
 	if (opt_list_mapcnt)
-		printf("%" PRIu64 "\t", mapcnt);
+		printf("%lu\t", mapcnt);
 
 	printf("%lx\t%s\n", offset, page_flag_name(flags));
 }
